refactor(ui): extract hero section from landing page into sub-components
Break the 2,379-line page.tsx monolith into focused components:
- hero-section.tsx: orchestrator (scroll, parallax, layout)
- hero-nav.tsx: navigation bar with hover indicator
- hero-sidebar.tsx: sidebar with history and wallet
- chat-messages.tsx: message list with scroll button
- chat-input.tsx: input form, dog illustration, wallet cards
